+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Autumnity Integration!
+  * Eating with the Foul Taste effect now increases heal by 50%
 * If player's stamina is below 20 (2 hearts) and stamina is locked, the player is slowed down by 25%
 * Knockback
   * Pickaxes, axes and ITE Forge Hammers now deal reduced knockback